Human vs Computer
Added code that allows the black side to be controlled by the computer.
Code is sloppy and is eventually going to be changed around but this at
least allows some basic AI testing to be done. Successive move logic
not yet complete. Need to add checks to see if a piece has gone to the
same location twice, and if it has traveled the same direction twice in
a row