Region specific redraw
Changed the window to only redraw the timer when a timer event fires, and redraw the window in the event of a click. A pause is used because calling clearWindow() at the end of createWindow without a delay does nothing. Adding the pause causes it to work correctly.

Also removed more parts that were not needed from the code.