Made game playable with the changes
Finished the major changes to the point where the game is playable
again (although I suspect it will crash when non capturing moves become
available again, more testing required)

In summary: Changed the input format to that specified in the latest
requirements. Added support for any (odd) sized board. Updated the
whiteMovesFull to reflect the changes in input format. Changed the way
the rows and columns are extracted from strings to allow for board
sizes with rows and columns with multiple digits.