Move selection implemented
Now allows for the selection of a move by clicking on a coordinate on the board and then clicking on a different coordinate. Currently the second point is shown with a blue selection box, and there is no checking of validity of moves.

Added the initial processMove function that will actually send the move to the Fanorona game to be handled.