Added more client server functionality
The game now plays more like it should. The server is computer
controlled and the client human. Will given an error on the second
human move, need to get to the bottom of that.

Also cleaned up the interface to get an possible move from the AI.
Basically just moved what was in main to Fanorona.