Added MiniMaxTree getPath
Adds a function to the MiniMax tree that will return in a linked list the game states from the root node to the best child node. This is supposed to be called after process tree has been called first.

Also updated the tests to work with some of the MiniMaxTree changes.