Don't process grid clicks without game
Fixed handling of clicks so that they are not interpreted in terms of the grid if the game is not visible.

Also hid the nearest coordinate information as it is not needed anymore.

Clicks on buttons cannot be interpreted as clicks on the board anymore.

The time is now hidden if the game is hidden.